+++ code/game/src/world_view.h
@@ -16,8 +16,8 @@ typedef struct {
     uint16_t chunk_amount;
     
     // NOTE(zaklaus): metrics
-    uint64_t last_update[WORLD_TRACKER_LAYERS];
-    uint64_t delta_time[WORLD_TRACKER_LAYERS];
+    float last_update[WORLD_TRACKER_LAYERS];
+    float delta_time[WORLD_TRACKER_LAYERS];
     uint8_t active_layer_id;
 } world_view;
